refactor: :technologist: refactored interactionCreate
Improved handling of interactions, using correct types
